break up odbapi
