+++ explore/src/main/scala/explore/services/OdbApi.scala
@@ -5,39 +5,10 @@ package explore.services
 
 import cats.MonadThrow
 import cats.effect.IO
-import explore.model.ObsIdSet
 import japgolly.scalajs.react.React
 import japgolly.scalajs.react.feature.Context
-import lucuma.core.model.Program
-import lucuma.core.model.Target
-import lucuma.schemas.ObservationDB.Enums.Existence
-import lucuma.schemas.ObservationDB.Types.UpdateTargetsInput
-import lucuma.schemas.model.TargetWithId
 
-trait OdbApi[F[+_]: MonadThrow]:
-  private val Uninitialized: F[Nothing] =
-    MonadThrow[F].raiseError:
-      new RuntimeException("OdbApi is not initialized.")
-
-  def insertTarget(programId: Program.Id, target: Target.Sidereal): F[Target.Id] = Uninitialized
-
-  def updateTarget(targetId: Target.Id, input: UpdateTargetsInput): F[Unit] = Uninitialized
-
-  def setTargetExistence(
-    programId: Program.Id,
-    targetId:  Target.Id,
-    existence: Existence
-  ): F[Unit] = Uninitialized
-
-  def deleteTargets(targetIds: List[Target.Id], programId: Program.Id): F[Unit] = Uninitialized
-
-  def undeleteTargets(targetIds: List[Target.Id], programId: Program.Id): F[Unit] = Uninitialized
-
-  def cloneTarget(
-    targetId:  Target.Id,
-    replaceIn: ObsIdSet,
-    input:     UpdateTargetsInput
-  ): F[TargetWithId] = Uninitialized
+trait OdbApi[F[+_]: MonadThrow] extends OdbTargetApi[F]
 
 object OdbApi:
   // Default value noop implementations with warning
